+++ eox_tagging/test/test_utils.py
@@ -1,6 +1,7 @@
 """
 Utils to run tests
 """
+from django.db import models
 from opaque_keys.edx.django.models import CourseKeyField
 
 try:
@@ -18,7 +19,8 @@ try:
 
     class CourseEnrollment(fake.FakeModel):
         """Fake Model enrollments."""
-        pass
+        username = models.CharField(max_length=255, null=True)
+        course_id = models.CharField(max_length=255, null=True)
 
 except ImportError:
     CourseEnrollment = object