+++ eox_tagging/urls.py
@@ -1,7 +1,8 @@
 """
 eox_tagging URL Configuration
 """
-from django.conf.urls import url  # pylint: disable=unused-import
-from django.contrib import admin  # pylint: disable=unused-import
+from django.conf.urls import include, url
 
-urlpatterns = []
+urlpatterns = [
+    url(r'api/', include('eox_tagging.api.urls')),
+]