+++ CHANGELOG.rst
@@ -6,11 +6,15 @@ Change Log
    in this file.  It adheres to the structure of http://keepachangelog.com/ ,
    but in reStructuredText instead of Markdown (for ease of incorporation into
    Sphinx documentation and the PyPI description).
-   
+
    This project adheres to Semantic Versioning (http://semver.org/).
 .. There should always be an "Unreleased" section for changes pending release.
+
 Unreleased
 ----------
+
+- REST API to create, get, filter and delete tags.
+- New filters in Tag queryset.
 
 * First PyPI release.